crawler/pyproject.toml
[project]
name = "quality-pulse-crawler"
version = "0.1.0"
description = ""
authors = [
    {name = "Seonguk Choi",email = "tksqkac@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crawl4ai (>=0.8.5,<0.9.0)",
    "playwright (>=1.58.0,<2.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.3.0,<8.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "google-genai (>=1.68.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
